// libs/shared/modules/src/data/models/domain/control.ts
/**
 * Anecdotes API
 * Anecdotes API's
 *
 * OpenAPI spec version: 0.0.3
 * Contact: apiteam@anecdotes.ai
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ControlStatus } from './controlStatus';
export interface Control { 
  /**
  * Anecdotes framework control name
  */
  control_name?: string;
  /**
   * the id of the underlying global control
   */
  control_id?: string;

  control_framework_id?: string;
  /**
   * The catgeory of the control (e.g. \"AC Access Control\")
   */
  control_category?: string;
  /**
   * List of frameworks ids related with this control
   */
  control_related_frameworks?: Array<string>;
  control_related_frameworks_names?: { [framework_name: string]: string[] };
  /**
   * Whether this control is applicable for the customer or not
   */
  control_is_applicable?: boolean;
  /**
   * Anecdotes framework control description
   */
  control_description?: string;
  /**
   * Whether this control is a custom control created by the user or not
   */
  control_is_custom?: boolean;
  /**
   * The user who last added/edited this control
   */
  control_edited_by?: string;
  /**
   * Determine whethr control has note
   */
  control_note_exists?: boolean;
  /**
   * The time this control was last added/edited
   */
  control_last_edit_time?: string;
  control_status?: ControlStatus;
  /**
   * The control implementation guidance for customers
   */
  control_guidance_for_customers?: string;
  /**
   * The control implementation guidance for providers
   */
  control_guidance_for_providers?: string;
  /**
   * The email of control's owner
   */
  control_owner?: string;

  control_requirement_ids?: string[];

  snapshot_id?: string;

  is_snapshot?: boolean;
}